Rewrote the foundation concepts intro.
This stuff in this section is heavy, but I think it is very important
to understand how things work. I know that for me personally, I didn’t
understand how to really make complex features until I wrote a layout
engine. That gave me an understanding of how everything worked on the
inside. Obviously not everyone is going to do that so I want to give
the essential knowledge.